test: storage - api v4 - fix tests and improve tests structure
